\name{alarm}
\alias{alarm}
\docType{data}

\title{One example dataset for \eqn{p}-plearning algorithm.} 

\description{
The ALARM ("A Logical Alarm Reduction Mechanism") is a Bayesian network designed to provide an alarm message system for patient monitoring.
\describe{
  \item{ alarm }{ An object of class 'bn' for 'alarm' dataset. See package \pkg{bnlearn} for detail.}
}
}
\usage{data(alarm)}
\format{
  Alarm dataset is an object of class \code{bn} or \code{bn.fit}. See package \pkg{bnlearn} for detail.
}


\references{
  
Xu, S., Jia, B., and Liang, F. (2018). Learning Moral Graphs in Construction of High-Dimensional Bayesian Networks for Mixed Data. Submitted. 
  
  
I. A. Beinlich, H. J. Suermondt, R. M. Chavez, and G. F. Cooper. The ALARM Monitoring System: A Case Study with Two Probabilistic Inference Techniques for Belief Networks. In Proceedings of the 2nd European Conference on Artificial Intelligence in Medicine, pages 247-256. Springer-Verlag, 1989.
}


\keyword{datasets}
