\name{count}
\alias{count}
\docType{data}

\title{An example of count dataset for constructing networks} 

\description{
count is a simulated dataset for illustrating our proposed method for inferening networks from next generation sequencing data.
}
\usage{data(count)}
\format{
 count dataset is a 100x200 matrix. Each row represents a observation and each column represents a variable. It is generated from an overdispersion and zero-inflated Possion distribution. 
}


\references{
Jia, B., Xu, S., Xiao, G., Lamba, V., Liang, F. (2017) Inference of Genetic Networks from Next Generation Sequencing Data. Biometrics, in press.
}


\keyword{datasets}
