\name{mixed3000}
\alias{mixed3000}
\docType{data}

\title{One example dataset for p_learning} 

\description{
  mixed3000 is a simulated dataset for illustration our p_learning alogorithm.
}
\usage{data(mixed3000)}
\format{
  mixed3000 dataset is a 3000x100 matrix. Each row represents a observation and each column represents a variable. It contains 50 continuous variables and 50 binary variables.  
}


\references{
  
  Suwa, Xu and Faming, Liang (2017). Learning High-Dimensional Bayesian Networks for General Types of
  Random Variables. Submitted to Biometrika. 
  
}


\keyword{datasets}
