\name{TR0}
\alias{TR0}
\docType{data}

\title{One example dataset for equSA} 

\description{
TR0 is a simulated dataset for illustration our equSA alogorithm.
}
\usage{data(TR0)}
\format{
 TR0 dataset is a 100x200 matrix. Each row represents a observation and each column represents a variable.   
}


\references{
 Liang, F., Song, Q. and Qiu, P. (2015). An Equivalent Measure of Partial Correlation Coefficients for High Dimensional Gaussian Graphical Models. J. Amer. Statist. Assoc., 110, 1248-1265.
}


\keyword{datasets}
