\name{TR0_mat}
\alias{TR0_mat}
\docType{data}

\title{The adjacency matrix for TR0 dataset.} 

\description{
TR0_mat is an estimated adjacency matrix by \eqn{\psi}- learning alogorithm.
}
\usage{data(TR0_mat)}
\format{
 TR0_mat a 200x200 matrix with binary values. When its element (i,j) equals to 1, there exists an edge between variable i and j. Otherwise, it equals to 0.
}


\references{
 Liang, F., Song, Q. and Qiu, P. (2015). An Equivalent Measure of Partial Correlation Coefficients for High Dimensional Gaussian Graphical Models. J. Amer. Statist. Assoc., 110, 1248-1265.
}


\keyword{datasets}
